[mod/#126] MyPage 구조 수정
